+++ src/api.deprecated.d.ts
@@ -1,0 +1,26 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import { Uri } from 'vscode';
+
+declare module './api' {
+    export interface JupyterServerCommandProvider {
+        /**
+         * Returns a list of commands to be displayed to the user.
+         * @deprecated Use `provideCommands` instead.
+         */
+        commands?: JupyterServerCommand[];
+    }
+
+    export interface JupyterServerCommand {
+        /**
+         * @deprecated Use `label` instead.
+         */
+        title?: string;
+        /**
+         * @deprecated Use `description` instead.
+         */
+        detail?: string;
+    }
+}
